+++ src/main/java/at/hannibal2/skyhanni/config/features/RiftConfig.java
@@ -11,9 +11,9 @@ public class RiftConfig {
     @ConfigOption(name = "Rift Timer", desc = "")
     @Accordion
     @Expose
-    public TimerConfig timer = new TimerConfig();
+    public RiftTimerConfig timer = new RiftTimerConfig();
 
-    public static class TimerConfig {
+    public static class RiftTimerConfig {
 
         @Expose
         @ConfigOption(name = "Enabled", desc = "Show the remaining rift time, max time, percentage, and extra time changes.")
@@ -34,4 +34,9 @@ public class RiftConfig {
         public Position timerPosition = new Position(10, 10, false, true);
 
     }
+
+    @Expose
+    @ConfigOption(name = "Highlight Guide", desc = "Highlight things to do in the Rift Guide.")
+    @ConfigEditorBoolean
+    public boolean highlightGuide = true;
 }